+++ pkg/tfgen/generate_schema.go
@@ -22,11 +22,13 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
+	"sort"
 	"strings"
 
 	"github.com/gedex/inflector"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/pkg/errors"
+	"github.com/pulumi/pulumi/pkg/v2/codegen"
 	pschema "github.com/pulumi/pulumi/pkg/v2/codegen/schema"
 	"github.com/pulumi/pulumi/sdk/v2/go/common/tokens"
 	"github.com/pulumi/pulumi/sdk/v2/go/common/util/contract"
@@ -74,8 +76,12 @@ func (g *schemaGenerator) typeName(r *resourceType) string {
 }
 
 type schemaNestedType struct {
-	typ       *propertyType
-	pyMapCase bool
+	typ             *propertyType
+	declarer        declarer
+	required        codegen.StringSet
+	requiredInputs  codegen.StringSet
+	requiredOutputs codegen.StringSet
+	pyMapCase       bool
 }
 
 type schemaNestedTypes struct {
@@ -103,16 +109,16 @@ func gatherSchemaNestedTypesForMember(member moduleMember) map[string]*schemaNes
 func (nt *schemaNestedTypes) gatherFromMember(member moduleMember) {
 	switch member := member.(type) {
 	case *resourceType:
-		nt.gatherFromProperties(member, member.name, "", member.inprops, true)
-		nt.gatherFromProperties(member, member.name, "", member.outprops, true)
+		nt.gatherFromProperties(member, member.name, member.inprops, true, true)
+		nt.gatherFromProperties(member, member.name, member.outprops, false, true)
 		if !member.IsProvider() {
-			nt.gatherFromProperties(member, member.name, "", member.statet.properties, true)
+			nt.gatherFromProperties(member, member.name, member.statet.properties, true, true)
 		}
 	case *resourceFunc:
-		nt.gatherFromProperties(member, member.name, "", member.args, true)
-		nt.gatherFromProperties(member, member.name, "", member.rets, true)
+		nt.gatherFromProperties(member, member.name, member.args, true, true)
+		nt.gatherFromProperties(member, member.name, member.rets, false, true)
 	case *variable:
-		nt.gatherFromPropertyType(member, member.name, "", "", member.typ, true)
+		nt.gatherFromPropertyType(member, member.name, "", member.typ, false, true)
 	}
 }
 
@@ -121,25 +127,61 @@ type declarer interface {
 }
 
 func (nt *schemaNestedTypes) declareType(
-	declarer declarer, namePrefix, name, nameSuffix string, typ *propertyType, pyMapCase bool) string {
+	declarer declarer, namePrefix, name string, typ *propertyType, isInput, pyMapCase bool) string {
 
 	// Generate a name for this nested type.
-	baseName := namePrefix + strings.Title(name)
+	typeName := namePrefix + strings.Title(name)
 
 	// Override the nested type name, if necessary.
 	if typ.nestedType.Name().String() != "" {
-		baseName = typ.nestedType.Name().String()
+		typeName = typ.nestedType.Name().String()
 	}
 
-	typeName := baseName + strings.Title(nameSuffix)
 	typ.name = typeName
 
-	nt.nameToType[typeName] = &schemaNestedType{typ: typ, pyMapCase: pyMapCase}
-	return baseName
+	required := codegen.StringSet{}
+	for _, p := range typ.properties {
+		if !p.optional() {
+			required.Add(p.name)
+		}
+	}
+
+	var requiredInputs, requiredOutputs codegen.StringSet
+	if isInput {
+		requiredInputs = required
+	} else {
+		requiredOutputs = required
+	}
+
+	if existing, ok := nt.nameToType[typeName]; ok {
+		contract.Assert(existing.declarer == declarer || existing.typ.equals(typ))
+
+		// For output type conflicts, record the output type's required properties. These will be attached to
+		// a nodejs-specific blob in the object type's spec s.t. the node code generator can generate code that matches
+		// the code produced by the old tfgen code generator.
+		if isInput {
+			existing.requiredInputs = requiredInputs
+		} else {
+			existing.requiredOutputs = requiredOutputs
+		}
+
+		existing.typ, existing.required = typ, required
+		return typeName
+	}
+
+	nt.nameToType[typeName] = &schemaNestedType{
+		typ:             typ,
+		declarer:        declarer,
+		required:        required,
+		requiredInputs:  requiredInputs,
+		requiredOutputs: requiredOutputs,
+		pyMapCase:       pyMapCase,
+	}
+	return typeName
 }
 
-func (nt *schemaNestedTypes) gatherFromProperties(declarer declarer, namePrefix, nameSuffix string, ps []*variable,
-	pyMapCase bool) {
+func (nt *schemaNestedTypes) gatherFromProperties(declarer declarer, namePrefix string, ps []*variable,
+	isInput, pyMapCase bool) {
 
 	for _, p := range ps {
 		name := p.name
@@ -151,21 +193,21 @@ func (nt *schemaNestedTypes) gatherFromProperties(declarer declarer, namePrefix,
 		// properties an object-typed element that are not Map types. This is consistent with the earlier behavior. See
 		// https://github.com/pulumi/pulumi/issues/3151 for more details.
 		mapCase := pyMapCase && p.typ.kind == kindObject && p.schema.Type == schema.TypeMap
-		nt.gatherFromPropertyType(declarer, namePrefix, name, nameSuffix, p.typ, mapCase)
+		nt.gatherFromPropertyType(declarer, namePrefix, name, p.typ, isInput, mapCase)
 	}
 }
 
-func (nt *schemaNestedTypes) gatherFromPropertyType(
-	declarer declarer, namePrefix, name, nameSuffix string, typ *propertyType, pyMapCase bool) {
+func (nt *schemaNestedTypes) gatherFromPropertyType(declarer declarer, namePrefix, name string, typ *propertyType,
+	isInput, pyMapCase bool) {
 
 	switch typ.kind {
 	case kindList, kindSet, kindMap:
 		if typ.element != nil {
-			nt.gatherFromPropertyType(declarer, namePrefix, name, nameSuffix, typ.element, pyMapCase)
+			nt.gatherFromPropertyType(declarer, namePrefix, name, typ.element, isInput, pyMapCase)
 		}
 	case kindObject:
-		baseName := nt.declareType(declarer, namePrefix, name, nameSuffix, typ, pyMapCase)
-		nt.gatherFromProperties(declarer, baseName, nameSuffix, typ.properties, pyMapCase)
+		baseName := nt.declareType(declarer, namePrefix, name, typ, isInput, pyMapCase)
+		nt.gatherFromProperties(declarer, baseName, typ.properties, isInput, pyMapCase)
 	}
 }
 
@@ -250,15 +292,22 @@ func (g *schemaGenerator) genPackageSpec(pack *pkg) (pschema.PackageSpec, error)
 		spec.Types[token] = typ
 	}
 
-	if jsi := g.info.JavaScript; jsi != nil {
-		spec.Language["nodejs"] = rawMessage(map[string]interface{}{
-			"packageName":        jsi.PackageName,
-			"packageDescription": generateManifestDescription(g.info),
-			"dependencies":       jsi.Dependencies,
-			"devDependencies":    jsi.DevDependencies,
-			"typescriptVersion":  jsi.TypeScriptVersion,
-		})
+	downstreamLicense := g.info.GetTFProviderLicense()
+	licenseTypeURL := getLicenseTypeURL(downstreamLicense)
+	nodeReadme := fmt.Sprintf(
+		standardDocReadme, g.pkg, g.info.Name, g.info.GetGitHubOrg(), downstreamLicense, licenseTypeURL)
+	nodeData := map[string]interface{}{
+		"readme":                  nodeReadme,
+		"disableUnionOutputTypes": true,
 	}
+	if jsi := g.info.JavaScript; jsi != nil {
+		nodeData["packageName"] = jsi.PackageName
+		nodeData["packageDescription"] = generateManifestDescription(g.info)
+		nodeData["dependencies"] = jsi.Dependencies
+		nodeData["devDependencies"] = jsi.DevDependencies
+		nodeData["typescriptVersion"] = jsi.TypeScriptVersion
+	}
+	spec.Language["nodejs"] = rawMessage(nodeData)
 
 	if pi := g.info.Python; pi != nil {
 		spec.Language["python"] = rawMessage(map[string]interface{}{
@@ -453,6 +502,18 @@ func (g *schemaGenerator) genDatasourceFunc(mod string, fun *resourceFunc) psche
 	return spec
 }
 
+func setEquals(a, b codegen.StringSet) bool {
+	if len(a) != len(b) {
+		return false
+	}
+	for k := range a {
+		if !b.Has(k) {
+			return false
+		}
+	}
+	return true
+}
+
 func (g *schemaGenerator) genObjectType(mod string, typInfo *schemaNestedType) (string, pschema.ObjectTypeSpec) {
 	typ := typInfo.typ
 	contract.Assert(typ.kind == kindObject)
@@ -480,7 +541,45 @@ func (g *schemaGenerator) genObjectType(mod string, typInfo *schemaNestedType) (
 		}
 	}
 
+	nodeInfo := map[string]interface{}{}
+	if !setEquals(typInfo.required, typInfo.requiredInputs) {
+		requiredInputs := make([]string, 0, len(typInfo.requiredInputs))
+		for name := range typInfo.requiredInputs {
+			requiredInputs = append(requiredInputs, name)
+		}
+		sort.Strings(requiredInputs)
+		nodeInfo["requiredInputs"] = requiredInputs
+	}
+	if !setEquals(typInfo.required, typInfo.requiredOutputs) {
+		requiredOutputs := make([]string, 0, len(typInfo.requiredOutputs))
+		for name := range typInfo.requiredOutputs {
+			requiredOutputs = append(requiredOutputs, name)
+		}
+		sort.Strings(requiredOutputs)
+		nodeInfo["requiredOutputs"] = requiredOutputs
+	}
+	if len(nodeInfo) != 0 {
+		spec.Language = map[string]json.RawMessage{
+			"nodejs": rawMessage(nodeInfo),
+		}
+	}
+
 	return token, spec
+}
+
+func (g *schemaGenerator) schemaPrimitiveType(k typeKind) string {
+	switch k {
+	case kindBool:
+		return "boolean"
+	case kindInt:
+		return "integer"
+	case kindFloat:
+		return "number"
+	case kindString:
+		return "string"
+	default:
+		return ""
+	}
 }
 
 func (g *schemaGenerator) schemaType(mod string, typ *propertyType, out bool) pschema.TypeSpec {
@@ -489,6 +588,9 @@ func (g *schemaGenerator) schemaType(mod string, typ *propertyType, out bool) ps
 	case typ == nil:
 		return pschema.TypeSpec{Ref: "pulumi.json#/Any"}
 	case typ.typ != "" || len(typ.altTypes) != 0:
+		// Compute the default type for the union. May be empty.
+		defaultType := g.schemaPrimitiveType(typ.kind)
+
 		var toks []tokens.Type
 		if typ.typ != "" {
 			toks = []tokens.Type{typ.typ}
@@ -507,16 +609,9 @@ func (g *schemaGenerator) schemaType(mod string, typ *propertyType, out bool) ps
 				if pkg == g.pkg {
 					pkg = ""
 				}
-				spec := pschema.TypeSpec{Ref: fmt.Sprintf("%s#/types/%s", pkg, strings.TrimSuffix(string(t), "[]"))}
-				switch typ.kind {
-				case kindBool:
-					spec.Type = "boolean"
-				case kindInt:
-					spec.Type = "integer"
-				case kindFloat:
-					spec.Type = "number"
-				case kindString:
-					spec.Type = "string"
+				spec := pschema.TypeSpec{
+					Type: defaultType,
+					Ref:  fmt.Sprintf("%s#/types/%s", pkg, strings.TrimSuffix(string(t), "[]")),
 				}
 				if strings.HasSuffix(string(t), "[]") {
 					items := spec
@@ -528,7 +623,10 @@ func (g *schemaGenerator) schemaType(mod string, typ *propertyType, out bool) ps
 		if len(typs) == 1 {
 			return typs[0]
 		}
-		return pschema.TypeSpec{OneOf: typs}
+		return pschema.TypeSpec{
+			Type:  defaultType,
+			OneOf: typs,
+		}
 	case typ.asset != nil:
 		if typ.asset.IsArchive() {
 			return pschema.TypeSpec{Ref: "pulumi.json#/Archive"}
@@ -538,14 +636,10 @@ func (g *schemaGenerator) schemaType(mod string, typ *propertyType, out bool) ps
 
 	// First figure out the raw type.
 	switch typ.kind {
-	case kindBool:
-		return pschema.TypeSpec{Type: "boolean"}
-	case kindInt:
-		return pschema.TypeSpec{Type: "integer"}
-	case kindFloat:
-		return pschema.TypeSpec{Type: "number"}
-	case kindString:
-		return pschema.TypeSpec{Type: "string"}
+	case kindBool, kindInt, kindFloat, kindString:
+		t := g.schemaPrimitiveType(typ.kind)
+		contract.Assert(t != "")
+		return pschema.TypeSpec{Type: t}
 	case kindSet, kindList:
 		items := g.schemaType(mod, typ.element, out)
 		return pschema.TypeSpec{Type: "array", Items: &items}